sysfs: move …device_get_time from boltd/device
