+++ boltd/bolt-sysfs.h
@@ -29,6 +29,15 @@ struct udev_device;
 
 G_BEGIN_DECLS
 
+typedef enum BoltStatTime {
+  BOLT_ST_ATIME,
+  BOLT_ST_CTIME,
+  BOLT_ST_MTIME
+} BoltStatTime;
+
+gint64               bolt_sysfs_device_get_time (struct udev_device *udev,
+                                                 BoltStatTime        st);
+
 gboolean             bolt_sysfs_device_is_domain (struct udev_device *udev);
 
 struct udev_device * bolt_sysfs_domain_for_device (struct udev_device *udev);
